- increased vamp cap bonus for Anti-Paladin from flat 15% to 100% - fixed vamp cap of all AP related skills to use this AP bonus - fixed boon setbit issue with bits higher than 32 - increased hellfire absorb vamp amount from 16% to 20% - increased battle ecstasy group vamp amount from 22.5% to 40% - increased battle ecstasy self vamp amount from 15% to 20% - increased Avernus arti proc damage from 150 to 200, still affected by soul and negative shield

+++ src/boon.c
@@ -3046,27 +3046,27 @@ void check_boon_completion(P_char ch, P_char victim, double data, int option)
 				if ((int)bdata.bonus == 1)
 				{
 					snprintf(buff, MAX_STRING_LENGTH, "%s", affected1_bits[(int)bdata.bonus2].flagLong);
-					af.bitvector = 1 << (int)bdata.bonus2;
+					af.bitvector = (long)1 << (int)bdata.bonus2;
 				}
 				if ((int)bdata.bonus == 2)
 				{
 					snprintf(buff, MAX_STRING_LENGTH, "%s", affected2_bits[(int)bdata.bonus2].flagLong);
-					af.bitvector2 = 1 << (int)bdata.bonus2;
+					af.bitvector2 = (long)1 << (int)bdata.bonus2;
 				}
 				if ((int)bdata.bonus == 3)
 				{
 					snprintf(buff, MAX_STRING_LENGTH, "%s", affected3_bits[(int)bdata.bonus2].flagLong);
-					af.bitvector3 = 1 << (int)bdata.bonus2;
+					af.bitvector3 = (long)1 << (int)bdata.bonus2;
 				}
 				if ((int)bdata.bonus == 4)
 				{
 					snprintf(buff, MAX_STRING_LENGTH, "%s", affected4_bits[(int)bdata.bonus2].flagLong);
-					af.bitvector4 = 1 << (int)bdata.bonus2;
+					af.bitvector4 = (long)1 << (int)bdata.bonus2;
 				}
 				if ((int)bdata.bonus == 5)
 				{
 					snprintf(buff, MAX_STRING_LENGTH, "%s", affected5_bits[(int)bdata.bonus2].flagLong);
-					af.bitvector5 = 1 << (int)bdata.bonus2;
+					af.bitvector5 = (long)1 << (int)bdata.bonus2;
 				}
 				if (!*buff)
 				{
